NXP-25633: Refactoring and cleaning without changing logic
Use a message for bucket id.
Externalize computation.
Misc cleaning.

+++ nuxeo-core/nuxeo-core-bulk/src/main/java/org/nuxeo/ecm/core/bulk/io/BulkJsonReader.java
@@ -30,8 +30,8 @@ import static org.nuxeo.ecm.core.io.registry.reflect.Priorities.REFERENCE;
 
 import java.time.Instant;
 
-import org.nuxeo.ecm.core.bulk.BulkStatus;
-import org.nuxeo.ecm.core.bulk.BulkStatus.State;
+import org.nuxeo.ecm.core.bulk.message.BulkStatus;
+import org.nuxeo.ecm.core.bulk.message.BulkStatus.State;
 import org.nuxeo.ecm.core.io.marshallers.json.EntityJsonReader;
 import org.nuxeo.ecm.core.io.registry.reflect.Setup;
 
@@ -67,12 +67,12 @@ public class BulkJsonReader extends EntityJsonReader<BulkStatus> {
 
         Long count = getLongField(jn, BULK_COUNT);
         if (count != null) {
-            status.setCount(count.longValue());
+            status.setCount(count);
         }
 
         Long processed = getLongField(jn, BULK_PROCESSED);
         if (processed != null) {
-            status.setProcessed(processed.longValue());
+            status.setProcessed(processed);
         }
 
         return status;
